move shaders out of loader
