+++ src/tiltbrush/tiltbrushshaders.d.ts
@@ -1,7 +1,4 @@
-declare const TiltBrushShaders: {
-    Light: {
-        Vert: any;
-        Frag: any;
-    };
-};
+import { ShaderMaterialParameters } from 'three';
+
+declare const TiltBrushShaders: { [name :string]: ShaderMaterialParameters };
 export { TiltBrushShaders };
